pyproject.toml
[tool.poetry]
name = "kasa"
version = "0.1.0"
description = ""
authors = ["TheBranchDriftCatalyst <djdanielsh@gmail.com>"]
readme = "README.md"
packages = [{include = "kasa_exporter"}]

[tool.poetry.dependencies]
python = "^3.12"
prometheus-client = "^0.20.0"
structlog = "^24.4.0"
python-kasa = "^0.7.0.5"
requests = "^2.32.3"
watchdog = "^4.0.1"
fastapi = "^0.111.1"
pytz = "^2024.1"
jsonpickle = "^3.2.2"
twisted = "^24.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
